Move SequenceManager to API
